+++ app/src/main/java/za/kilowatch/ultimatefilemanager/settings/IconCustomizationActivity.kt
@@ -508,7 +508,9 @@ class IconCustomizationActivity : AppCompatActivity() {
             IconItemData("toolbar_image_compress", getString(R.string.icon_toolbar_image_compress), R.drawable.ic_compress_image, emptyList()),
             IconItemData("toolbar_delete", getString(R.string.icon_toolbar_delete), R.drawable.ic_delete, emptyList()),
             IconItemData("toolbar_pin", getString(R.string.icon_toolbar_pin), R.drawable.ic_paperclip, emptyList()),
-            IconItemData("toolbar_unpin", getString(R.string.icon_toolbar_unpin), R.drawable.ic_paperclip_off, emptyList())
+            IconItemData("toolbar_unpin", getString(R.string.icon_toolbar_unpin), R.drawable.ic_paperclip_off, emptyList()),
+            IconItemData("toolbar_set_home_wallpaper", getString(R.string.icon_toolbar_set_home_wallpaper), R.drawable.ic_wallpaper_home, emptyList()),
+            IconItemData("toolbar_set_lock_wallpaper", getString(R.string.icon_toolbar_set_lock_wallpaper), R.drawable.ic_wallpaper_lock, emptyList())
         )
     }