+++ app/src/main/java/za/kilowatch/ultimatefilemanager/settings/ToolbarIconsActivity.kt
@@ -90,7 +90,9 @@ class ToolbarIconsActivity : AppCompatActivity() {
             IconItem(R.drawable.ic_paperclip, R.string.pin, ToolbarIconsPreferenceManager.KEY_PIN),
             IconItem(R.drawable.ic_paperclip_off, R.string.unpin, ToolbarIconsPreferenceManager.KEY_UNPIN),
             IconItem(R.drawable.ic_duplicate_finder, R.string.action_duplicate_finder, ToolbarIconsPreferenceManager.KEY_DUPLICATE_FINDER),
-            IconItem(R.drawable.ic_folder_large_files, R.string.action_large_files_finder, ToolbarIconsPreferenceManager.KEY_LARGE_FILES_FINDER)
+            IconItem(R.drawable.ic_folder_large_files, R.string.action_large_files_finder, ToolbarIconsPreferenceManager.KEY_LARGE_FILES_FINDER),
+            IconItem(R.drawable.ic_wallpaper_home, R.string.action_set_home_wallpaper, ToolbarIconsPreferenceManager.KEY_SET_HOME_WALLPAPER),
+            IconItem(R.drawable.ic_wallpaper_lock, R.string.action_set_lock_wallpaper, ToolbarIconsPreferenceManager.KEY_SET_LOCK_WALLPAPER)
         )