+++ app/src/main/java/za/kilowatch/ultimatefilemanager/settings/ToolbarIconsPreferenceManager.kt
@@ -27,6 +27,8 @@ object ToolbarIconsPreferenceManager {
     const val KEY_UNPIN = "pref_icon_unpin"
     const val KEY_DUPLICATE_FINDER = "pref_icon_duplicate_finder"
     const val KEY_LARGE_FILES_FINDER = "pref_icon_large_files_finder"
+    const val KEY_SET_HOME_WALLPAPER = "pref_icon_set_home_wallpaper"
+    const val KEY_SET_LOCK_WALLPAPER = "pref_icon_set_lock_wallpaper"
 
 
     fun isIconEnabled(context: Context, key: String): Boolean {